+++ CHANGELOG.md
@@ -3,3 +3,4 @@ This is the first public release of cls-scala-presentation-play-git.
 The most notable changes to prior (internal development) versions are:
 - The namespace is now `org.combinators`.
 - The project has been extracted to be standalone.
+- The shared routes `shared.routes` are now available as `org.combinators.cls.git.Routes`.